modifying test adj endpoint and integrating into ETL
