+++ app/namespaces/test_adjustment/__init__.py
@@ -1,1 +1,2 @@
 from .test_adjustment_controller import test_adjustment_ns
+from .test_adjustment_service import TestAdjHandler
